% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fences_df}
\alias{fences_df}
\title{Example Dataframe to Use with Fencing Algorithm}
\format{
Dataframe with 10 rows and 5 variables.
\describe{
\item{X}{x values.}
\item{Y}{y values.}
\item{prune_id}{prune id for fence, NA_character or oob_fence.}
\item{is_oob}{FALSE if at or inside the fence, TRUE otherwise
(i.e. out-of-bound).}
\item{label}{Label. Usually used for plotting.}
}
}
\usage{
fences_df
}
\description{
Example dataframe to use with fencing algorithm.
}
\details{
An example with data points outside of the \emph{fence} and \code{NA} that
are processed with the fencing algorithm. See \code{prune_fence} function
for details.
}
\keyword{datasets}
