% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_pruned}
\alias{iris_pruned}
\title{Example of Pruned Data with Iris}
\format{
Dataframe with 150 rows and 6 variables.
\describe{
\item{Sepal.Length}{Sepal length.}
\item{Sepal.Width}{Sepal width.}
\item{Petal.Length}{Petal length.}
\item{Petal.Width}{Petal width.}
\item{Species}{Species are iris setosa, versicolor and virginica.}
\item{prune_id}{prune id for fence, NA_character or oob_fence.}
}
}
\usage{
iris_pruned
}
\description{
Example of pruned data with iris.
}
\details{
See the \code{prune_fence} for more details. See \code{iris} for more
details on the original data.
}
\keyword{datasets}
