% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-A_prune.R
\name{prune}
\alias{prune}
\title{Identify Rows for Prunning}
\usage{
prune(data, id = NA_character_, prune_var = "prune_id", func = NULL, cols, ...)
}
\arguments{
\item{data}{Dataframe.}

\item{id}{String. ID used to identify the rows in column \code{prune_var}.}

\item{prune_var}{The name of the column with the prune id. Default is
"prune_id".}

\item{func}{Function used to select the rows meeting the condition for
prunning. Default is \code{NULL}. If \code{identity()} is used, nothing will
be done.}

\item{cols}{Names of columns to apply \code{func} to. Must be a character
vector with a minimum length of one.}

\item{...}{Additional arguments used by \code{func}.}
}
\value{
Dataframe with updated \code{prune_var} column.
}
\description{
Identify rows for prunning.
}
\details{
Set the column \code{prune_var} to \code{id} if the condition defined by
\code{func} is met.
}
\section{prune}{


When \code{func = NULL} and \code{id = NA_character_} and the column does not
exist, then it will be created and filled with \code{NA}. If the column
the column \code{prune_var} already exists, an error will be issued. If you
intend to reset the column values to \code{NA}, just delete it and create

When \code{func = NULL} and \code{id != NA_character}, then an error message
will be issued as an \code{id} cannot exist without a corresponding function.

When \code{func != NULL} and \code{id != NA_character}, then all elements
of column \code{prune_var} selected by \code{func} will be set to \code{id}
\emph{if they are NA or already set to id}, that is the rows with a
pre-existing value other than \code{id} will be kept as is.  This is
necessary to ensure that previous id are kept which is very useful when some
id are subsets of others be must be kept. For example when one wants to
keep zeros be exclude other values in a range then we would prune for zeros
\emph{then prune for the range}.

When \code{func != NULL} and \code{id = NA_character}, then an error message
will be issued as a function must always have a corresponding \code{id}.
}

\seealso{
prune_proc
}
