% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_03-B_divide.R
\name{prune_divide}
\alias{prune_divide}
\title{Divide 2 Variables Using Pruned Data}
\usage{
prune_divide(data, new_var, num_var, den_var, prune_var = "prune_id")
}
\arguments{
\item{data}{Dataframe.}

\item{new_var}{Name of new variable with the result of the computation.}

\item{num_var}{Name of numerator variable.}

\item{den_var}{Name of denominator variable.}

\item{prune_var}{The name of the column with the prune id. Default is
"prune_id".}
}
\value{
Data.frame.
}
\description{
Divide 2 variables using pruned data.
}
\details{
Divide 2 variables using pruned data. When the row has a prune id of
\code{NA}, the computation returns \code{NA}.
}
