% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_02-B_fence.R
\name{prune_fence}
\alias{prune_fence}
\title{Identify the Rows Outside the Fence.}
\source{
\emph{Statistical Data Cleaning}, Mark van der Loo and
Edwin de Jonge, 2018. Section 7.5.2, p. 176-179.
}
\usage{
prune_fence(data, cols, is_offset = TRUE)
}
\arguments{
\item{data}{Dataframe.}

\item{cols}{Names of columns to apply \code{func} to. Must be a character
vector with a minimum length of one.}

\item{is_offset}{If TRUE (default) the offset number will be
\code{offset = min(min(x), min(y))}, otherwise there will be not offset,
that is \code{offset = 0}.}
}
\value{
Logical vector where TRUE indicates values outside the fence.
}
\description{
Identify the rows outside the fence.
}
\details{
Compute the fence to eliminate values that are clearly out-of-bound.
Normally all values should be nonnegative. In case they are not, and
offset is used. Also, sometimes the data is nowhere near zero and is such
cases the fence is not useful, again in that case the offset solves that
problem.
}
