% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_03-E_fence.R
\name{prune_fence_df}
\alias{prune_fence_df}
\title{Update the fence data for pruning}
\usage{
prune_fence_df(df, x, y, specs)
}
\arguments{
\item{df}{Dataframe}

\item{x}{String, name of x variable}

\item{y}{String, name of y variable}

\item{specs}{Numeric vector of fence slopes and scale constant}
}
\value{
Dataframe with fence results
}
\description{
Update the fence data for pruning
}
\details{
Update the fence data for pruning using the slope calculated with
\code{prune_fence_slopes}
}
