% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_02-B_fence.R
\name{prune_fence_slopes}
\alias{prune_fence_slopes}
\title{Compute the Slopes Used for Fencing}
\usage{
prune_fence_slopes(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric vector. Must have finite values.}

\item{y}{Numeric vector. Must have finite values.}

\item{tol}{Number, tolerance when testing the ratios.}
}
\value{
List with \code{small_slope} and \code{big_slope}.
}
\description{
Compute the slopes used for fencing.
}
\details{
The function compute the slopes and also applies several tests to the result.
Problem with the data will always usually affect the ratio.
}
