% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_03-E_fence.R
\name{prune_fence_slopes_old}
\alias{prune_fence_slopes_old}
\title{Compute the fences' slope}
\usage{
prune_fence_slopes_old(x, y)
}
\arguments{
\item{x}{Numeric with only finite values}

\item{y}{Numeric with only finite values}
}
\value{
Numeric vectors with the 2 slopes
}
\description{
Compute the fences' slope
}
\details{
Compute the fences' slopes.
}
