% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-A_upd_reset.R
\name{prune_flag}
\alias{prune_flag}
\title{Update the prune flag}
\usage{
prune_flag(df, prune_var = "prune_id", flag_var = "is_pruned")
}
\arguments{
\item{df}{dataframe}

\item{prune_var}{String, name of the column with the prune id}

\item{flag_var}{String, name of the column with the prune flag}
}
\value{
dataframe with update column \code{prune_flag}
}
\description{
Update the prune flag.
}
\details{
Update the prune flag by setting it to TRUE when the string in
\code{prune_var} is not equal to \code{NA_character_} and set it to FALSE
otherwise.
}
\examples{
df <- data.frame(
 alpha = letters[1:3],
 prune_id = factor(c(NA_character_, "out", NA_character_)),
 is_pruned = FALSE)
df <- prune_flag(df)
# this should give the following updated dataframe
target <- data.frame(
 alpha = letters[1:3],
 prune_id = factor(c(NA_character_, "out", NA_character_)),
 is_pruned = c(FALSE, TRUE, FALSE))
stopifnot(identical(df, target))
}
