% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-C_oob.R
\name{prune_oob}
\alias{prune_oob}
\title{Prune Rows with Values Outside a Given Range}
\usage{
prune_oob(df, var, prune_var = "prune_id", id = "oob", rng = c(-Inf, Inf))
}
\arguments{
\item{df}{dataframe}

\item{var}{String, name of the numeric column to prune}

\item{prune_var}{String, name of the column with the prune id}

\item{id}{String, identification of the pruning in \code{prune_var}}

\item{rng}{Numeric, The range of values}
}
\value{
dataframe with updated \code{prune_var}  column
}
\description{
Prune rows with values outside a given range.
}
\details{
Flag the data lines with numeric values outside of the allowed
range set by the user. The "flag" column \code{prune_var} will be
updated to TRUE when the value tested is outside the range. The NA and Inf
values are treated as outside the range.
}
\examples{
rng <- c(4, 7)
df <- data.frame(iris, prune_id = factor(NA_character_))
df <- prune_oob(df, var = "Sepal.Length", rng = rng)
# This should give 12 pruned points with prune id == "range"
check <- sum(iris$Sepal.Length < rng[1] | iris$Sepal.Length > rng[2])
stopifnot(sum(!is.na(df$prune_id)) == check)
}
