% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_02-A_oob.R
\name{prune_oob_num}
\alias{prune_oob_num}
\title{Flag Rows Outside of a Given Range}
\usage{
prune_oob_num(data, cols, rng = c(-Inf, Inf))
}
\arguments{
\item{data}{Dataframe.}

\item{cols}{Names of columns to apply \code{func} to. Must be a character
vector with a minimum length of one.}

\item{rng}{The range of numerical values. The values outside of \code{rng}
will be identified with TRUE.}
}
\value{
Logical vector where TRUE indicates values outside of \code{rng}.
}
\description{
Flag Rows Outside of a Given Range
}
\seealso{
prune_proc
}
