% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-B_proc.R
\name{prune_proc}
\alias{prune_proc}
\title{Identify Rows for Prunning}
\usage{
prune_proc(
  data,
  id = NA_character_,
  prune_var = "prune_id",
  func = NULL,
  cols,
  ...
)
}
\arguments{
\item{data}{Dataframe.}

\item{id}{String. ID used to identify the rows in column \code{prune_var}.}

\item{prune_var}{The name of the column with the prune id. Default is
"prune_id".}

\item{func}{Function used to select the rows meeting the condition for
prunning. Default is \code{NULL}. If \code{identity()} is used, nothing will
be done.}

\item{cols}{Names of columns to apply \code{func} to. Must be a character
vector with a minimum length of one.}

\item{...}{Additional arguments used by \code{func}.}
}
\value{
Data.frame with updated \code{prune_var} column.
}
\description{
Identify rows for prunning.
}
\details{
Set the column \code{prune_var} to \code{id} if the condition defined by
\code{func} is met. See more details in the details section of \code{prune}.
}
\seealso{
prune
}
