% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_02-A_trans.R
\name{prune_trans}
\alias{prune_trans}
\title{Transform pruned data with \code{bestNormalize}}
\usage{
prune_trans(
  df,
  raw_var,
  trans_var,
  prune_var = "prune_id",
  fnc_trans = NULL,
  details = FALSE
)
}
\arguments{
\item{df}{dataframe}

\item{raw_var}{String, name of column to process}

\item{trans_var}{String, name of the column with transformed data}

\item{prune_var}{String, name of the column with the prune identifications}

\item{fnc_trans}{Function, created with the \code{bestNormalize}
package. If NULL, the default function \code{bestNormalize::no_transform}
is used.}

\item{details}{TRUE = create a list with the data and \code{bestNormalize}
object.}
}
\value{
dataframe with transformed values
}
\description{
Transform pruned data with \code{bestNormalize}.
}
\details{
Transform pruned data where the \code{prune_var == NA} using the
\code{bestNormalize} package.
}
\examples{
# the example dataframe
df <- data.frame(
 num = rlnorm(n = 5, meanlog = 1, sdlog = 2),
 prune_id = factor(NA_character_))
# Create the transformation function using the `bestNormalize` package
fnc_trans <- function(x) {
 bestNormalize::log_x(x, b = exp(1), standardize = FALSE)
}
# transform the data
df <- prune_trans(df, raw_var = "num", trans_var = "num_log",
 fnc_trans = fnc_trans, details = TRUE)
}
