% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_03-A_transform.R
\name{prune_transform}
\alias{prune_transform}
\title{Transform a Variable Using Pruned Data}
\usage{
prune_transform(data, var, prune_var = "prune_id", func = log, ...)
}
\arguments{
\item{data}{Dataframe.}

\item{var}{Name of variable with the result of the computation.}

\item{prune_var}{The name of the column with the prune id. Default is
"prune_id".}

\item{func}{Function used to transform \code{var}. Most common example
is \code{log}.}

\item{...}{Additional arguments used by \code{func}.}
}
\value{
Data.frame.
}
\description{
Transform a variable using pruned data.
}
\details{
Transform a variable using pruned data. When the row has a prune id of
\code{NA}, the computation returns \code{NA}.
}
