% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-A_upd_reset.R
\name{prune_upd}
\alias{prune_upd}
\title{Update the prune id}
\usage{
prune_upd(df, cond, prune_var = "prune_id", id)
}
\arguments{
\item{df}{dataframe.}

\item{cond}{logical vector where TRUE = update the id when it is not already
set, that is if its value ids not \code{id}.}

\item{prune_var}{String, name of the column with the prune id}

\item{id}{String, identification of the pruning in \code{prune_var}.
It must be different than  \code{NA_character_}.}
}
\value{
dataframe with update column \code{prune_var}.
}
\description{
Update the prune id
}
\details{
Set the prune id based on a logical vector to \code{id}
}
\examples{
df <- data.frame(
 alpha = letters[1:3],
 prune_id = factor(c(NA_character_, "out", NA_character_))
 )
df <- prune_upd(df, cond = c(FALSE, TRUE, TRUE), id = "example")
# this should update the third element to "example"
stopifnot(all.equal(df$prune_id, factor(c(NA_character_, "out", "example")),
 check.attributes = FALSE))
}
