% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_boxplot.R
\name{summ_boxplot}
\alias{summ_boxplot}
\title{Summary with \code{boxplot.stats}}
\source{
\url{https://stackoverflow.com/questions/56669653/boxplot-stats-in-dplyr-with-groups}
}
\usage{
summ_boxplot(
  data,
  box_var = "ratio_log",
  group_var = NULL,
  prune_var = NULL,
  log = TRUE,
  coef = 1.5
)
}
\arguments{
\item{data}{Dataframe.}

\item{box_var}{Name of variable used by \code{boxplot.stats}.}

\item{group_var}{Name of grouping variable. If \code{NULL}, the summary will
only have one line for all rows, without grouping.}

\item{prune_var}{Name of variable with prune id. If \code{NULL}, no filter
is used to remove pruned rows.}

\item{log}{TRUE: Convert computations with \code{exp}.}

\item{coef}{Numeric constant to determine how far the whisker extends. If
zero, the whiskers will extend to extremes and no outlier will be returned.
See \code{grDevices::boxplot.stats} for more details.}
}
\value{
Data.frame with boxplot.stats by \code{group_var}.
}
\description{
Summary with \code{boxplot.stats}.
}
\details{
The summary allows to exclude pruned data from the stats.
}
\seealso{
grDevices boxplot.stats
}
