% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_pruned.R
\name{summ_pruned}
\alias{summ_pruned}
\title{Summary of Pruned Rows}
\usage{
summ_pruned(data, prune_var = "prune_id", amt_var = NULL)
}
\arguments{
\item{data}{Data.frame.}

\item{prune_var}{Name of variable with prune id. Default value is "prune_id".}

\item{amt_var}{Name of optional variable to sum. If NULL (default) then no
sum of amount will appear in the summary.}
}
\value{
Dataframe with summary info.
}
\description{
Summary of pruned rows.
}
\details{
Summary of pruned rows with info on the number of rows and the sum of
an optional variable identified by \code{amt_var},
}
