% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_03-E_fence.R
\name{prune_fence}
\alias{prune_fence}
\title{Flag the data lines outside the fence}
\source{
Statistical Data Cleaning by Mark van der Loo and Edwin de Jonge
in section 7.5.2, p. 177-179.
}
\usage{
prune_fence(df, x, y, prune_var = "prune_id", id = "fence", details = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{x}{String, name of the numeric column "X" to test}

\item{y}{String, name of the numeric column "Y" to test.}

\item{prune_var}{String, name of the column with the prune identifications}

\item{id}{String, identification of the pruning in \code{prune_var}}

\item{details}{Flag, TRUE return a list with extra info on slopes and
data supporting the fence calculations. Default is \code{FALSE}.}
}
\value{
dataframe with an extra flag column
}
\description{
prune data outside the fence.
}
\details{
Prune data within the outside the fence. Rows with \code{prune_id} not NA
or with non-finite values will be excluded from the fence.
Normally, data should be pruned in the following order
\itemize{
\item non-finite: Lines with \code{Inf} and \code{NA}
\item Out-of-bound: Lines outside of reasonable range set by user
\item Fence: Lines outside the fences delimited by the slopes
\code{min(x) / max(y)} and \code{max(x) / min(y)}.
See details and the source.
\emph{Important:} Fencing is not always a good thing. For example with a data
set where the data is contains in an elongated area, e.g.
the iris data set for Sepal.Width and Petal.Width.
}
}
\section{The Fence}{


The fence is an algebraic way of testing the proposition
if x > A then y > 0 AND if y > B then x > 0. See the 2 tests with plots to
illustrate in test scripts.
}

\examples{
df <- data.frame(
 valX = c(0.0, 0.0, 0.3, 1.0, 3.0, 3.0, 3.0, 7.5, 8.0, 8.2),
 valY = c(0.0, 0.5, 6.0, 2.0, 0.0, 2.5, 8.1, 1.0, 7.0, 6.0),
 prune_id = factor(NA_character_)
 )
df <- prune_fence(df, x = "valX", y = "valY")
# This should give pruned 3 points
stopifnot(sum(!is.na(df$prune_id)) == 3)
}
