% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_01-A_upd_reset.R
\name{prune_reset}
\alias{prune_reset}
\title{Reset the prune id column}
\usage{
prune_reset(df, prune_var = "prune_id", id = NA_character_)
}
\arguments{
\item{df}{dataframe}

\item{prune_var}{String, name of the column with the prune id}

\item{id}{String, if not NA then only that id will be reset to \code{NA}}
}
\value{
dataframe with reset \code{prune_var}
}
\description{
Reset the prune id column.
}
\details{
If it doesn't already exists it will create the prune id column. The id
will be reset to \code{NA_character_} in all cases.
}
\examples{
df <- data.frame(alpha = letters[1:3])
df <- prune_reset(df, prune_var = "prune_id")
}
