% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_boxplot.R
\name{summ_boxplot}
\alias{summ_boxplot}
\title{Summary with \code{boxplot.stats}}
\usage{
summ_boxplot(
  data,
  group_var = "group",
  box_var = "ratio_log",
  prune_var = NULL,
  log = TRUE
)
}
\arguments{
\item{data}{Dataframe.}

\item{group_var}{Name of grouping variable.}

\item{box_var}{Name of variable used by \code{boxplot.stats}.}

\item{prune_var}{Name of variable with prune id. If \code{NULL}, no filter
is used to remove pruned rows.}

\item{log}{TRUE: Convert computations with \code{exp}.}
}
\value{
Data.frame with boxplot.stats by \code{group_var}.
}
\description{
Summary with \code{boxplot.stats}.
}
\details{
The summary allows to exclude pruned data from the stats.
}
