% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats.R
\name{summ_stats}
\alias{summ_stats}
\title{Summary with \code{n, sum}}
\usage{
summ_stats(
  data,
  group_var = "group",
  stats_var = "ratio_log",
  prune_var = NULL,
  log = TRUE
)
}
\arguments{
\item{data}{Dataframe.}

\item{group_var}{Name of grouping variable.}

\item{stats_var}{Name of variable used to compute the stats.}

\item{prune_var}{Name of variable with prune id. If \code{NULL}, no filter
is used to remove pruned rows.}

\item{log}{TRUE: Convert computations with \code{exp}.}
}
\value{
Data.frame with boxplot.stats by \code{group_var}.
}
\description{
Summary with \code{n, sum}.
}
\details{
The summary allows to exclude pruned data from the stats.
}
